requirements.txt
amqp==5.3.1
apturl==0.5.2
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==3.2.0
beautifulsoup4==4.13.5
billiard==4.2.2
blinker==1.4
branca==0.8.1
Brlapi==0.8.3
Brotli==1.1.0
cachetools==5.0.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
chardet==4.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
convertisseur==0.1.0
coverage==7.10.4
create-app-savane==0.0.4
cron_descriptor==2.0.6
cryptography==46.0.1
cssselect2==0.8.0
cupshelpers==1.0
daphne==4.2.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-leaflet==0.32.0
django-timezone-field==7.1
django-weasyprint==2.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djstruct-savane==0.0.1
dlib==20.0.0
dotenv==0.9.9
dotenv-wizard-savane==0.1.0
drf-yasg==1.21.10
duplicity==0.8.21
easyocr==1.7.2
et_xmlfile==2.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
filelock==3.19.1
flower==2.0.1
folium==0.17.0
fonttools==4.59.1
fsspec==2025.9.0
future==0.18.2
GDAL==3.4.1
geographiclib==2.1
geopy==2.4.1
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
gunicorn==23.0.0
httplib2==0.20.2
humanize==4.13.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
incremental==24.7.2
inflection==0.5.1
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==23.5.0
kombu==5.5.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
lxml==6.0.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
paramiko==2.9.3
pexpect==4.8.0
pillow==11.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==3.12.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pyclipper==1.3.0.post6
pycparser==2.23
pycups==2.0.1
PyDrive==1.3.1
pydyf==0.11.0
PyGObject==3.42.1
PyJWT==2.10.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==25.3.0
pyparsing==2.4.7
pyphen==0.17.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4.1
python-bidi==0.6.6
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pyxdg==0.27
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.5
rpds-py==0.27.1
rsa==4.8
scikit-image==0.25.2
scipy==1.15.3
SecretStorage==3.3.1
service-identity==24.2.0
shapely==2.1.1
six==1.16.0
soupsieve==2.8
SQLAlchemy==1.4.31
sqlparse==0.5.3
swagger-spec-validator==3.0.4
sympy==1.14.0
systemd-python==234
tifffile==2025.5.10
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.3.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
triton==3.4.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.5.0
usb-creator==0.3.7
vboxapi==1.0
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.10.0
xdg==5
xkit==0.0.0
xyzservices==2025.4.0
zipp==1.0.0
zope.interface==8.0
zopfli==0.2.3.post1
